9월2일

+++ 01_variable/src/edu/kh/variable/practice/CastingPractice1.java
@@ -15,7 +15,7 @@ public class CastingPractice1 {
 		System.out.println(iNum2 * dNum); //10.0 자동형변환
 		System.out.println((double)iNum1); //10.0
 		
-		System.out.println(iNum1 / (float)iNum2); //2.5
+		System.out.println(iNum1 / (double)iNum2); //2.5
 		System.out.println(dNum); //2.5
 		
 		System.out.println((int)fNum); //3 강제형변환
